Organize the command modules
